requirements.txt
# Web scraping
playwright==1.48.0
aiohttp==3.10.10

# Data handling
dataclasses-json==0.6.7

# Logging and utilities
python-dotenv==1.0.1
